refactor: 리뷰 미리보기 생성 객체 도출
